+++ src/components/layout/Sidebar.svelte
@@ -4,6 +4,7 @@
 		Edit2Icon,
 		DatabaseIcon,
 		CoffeeIcon,
+		ActivityIcon,
 		BookIcon,
 		BookOpenIcon,
 		SettingsIcon,
@@ -29,6 +30,11 @@
 			</a>
 		</li>
 		<li>
+			<a href="#analytics" class:active={$activePage === Page.analytics}>
+				<ActivityIcon class="icon-in-text" /> Analytics
+			</a>
+		</li>
+		<li>
 			<a href="#history" class:active={$activePage === Page.history}>
 				{#if $activePage === Page.history}
 					<BookOpenIcon class="icon-in-text" />
@@ -50,8 +56,6 @@
 				<CoffeeIcon class="icon-in-text" /> About
 			</a>
 		</li>
-		<!-- svelte-ignore missing-declaration -->
-		<li><span class="version">Version: <b>{__APP_VERSION__}</b></span></li>
 	</div>
 </ul>
 
@@ -67,8 +71,7 @@
 
 		background-color: var(--clr-bg2);
 
-		li a,
-		li .version {
+		li a {
 			$margin: 0.4rem;
 
 			display: block;
@@ -97,12 +100,6 @@
 				background-color: var(--clr-bg3);
 				color: var(--clr-accent-light);
 			}
-		}
-
-		li .version {
-			text-align: center;
-			cursor: default;
-			font-size: 0.8em;
 		}
 	}
 </style>