+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Releases
 
-The latest release can be found [here](https://github.com/tametsi/ligas/releases/latest).
+The latest version can be found [here](https://ligas.dev).
 
 ### [v1.0.0](https://github.com/tametsi/ligas/releases/tag/v1.0.0)
 
@@ -16,3 +16,14 @@ Hello World!
 - history recording (in `localStorage`)
 - layout toggle for runners on Edit-page
 - various fixes and improvements
+
+### [v1.2.0](https://github.com/tametsi/ligas/tree/v1.2.0)
+
+#### Changes
+
+- fix: invalid runner list in some cases ([#6](https://github.com/tametsi/ligas/issues/6))
+- add PWA support on [ligas.dev](https://ligas.dev)
+- move version string to about page
+- add analytics ([#4](https://github.com/tametsi/ligas/pull/4))
+
+**Full Changelog**: [`v1.1.0...v1.2.0`](https://github.com/tametsi/ligas/compare/v1.1.0...v1.2.0)